+++ src/test/kotlin/step3_racingcar/InputValidatorTest.kt
@@ -5,6 +5,7 @@ import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.should
 import io.kotest.matchers.string.startWith
+import step3_racingcar.validator.InputValidator
 
 class InputValidatorTest : BehaviorSpec({
     val inputValidator = InputValidator()
